Flyttet hentSoknad fra mappeService til soknadservice Flyttet debugendepunkter i journalpost til egen klasse, lagt til drift tags i openapi Fjernet behandlingsAar logik, returnerer alltid OK og iÅr. Byttet fra deprecated Validator for Omsorgspenger kroniskt sykt barn til ny variant. Refaktorering pepclient, lagt på Test/Lokalprofil på alltidtilgangclient
