+++ src/main/kotlin/no/nav/k9punsj/journalpost/JournalpostDriftRoutes.kt
@@ -1,0 +1,258 @@
+package no.nav.k9punsj.journalpost
+
+import kotlinx.coroutines.reactive.awaitFirst
+import net.logstash.logback.argument.StructuredArguments.keyValue
+import no.nav.k9punsj.RequestContext
+import no.nav.k9punsj.SaksbehandlerRoutes
+import no.nav.k9punsj.akjonspunkter.AksjonspunktKode
+import no.nav.k9punsj.akjonspunkter.AksjonspunktService
+import no.nav.k9punsj.akjonspunkter.AksjonspunktStatus
+import no.nav.k9punsj.felles.FagsakYtelseType
+import no.nav.k9punsj.felles.IdentOgJournalpost
+import no.nav.k9punsj.felles.Identitetsnummer.Companion.somIdentitetsnummer
+import no.nav.k9punsj.felles.IkkeFunnet
+import no.nav.k9punsj.felles.IkkeStøttetJournalpost
+import no.nav.k9punsj.felles.IkkeTilgang
+import no.nav.k9punsj.fordel.PunsjInnsendingType
+import no.nav.k9punsj.innsending.InnsendingClient
+import no.nav.k9punsj.integrasjoner.dokarkiv.SafDtos
+import no.nav.k9punsj.integrasjoner.gosys.GosysService
+import no.nav.k9punsj.integrasjoner.pdl.PdlService
+import no.nav.k9punsj.journalpost.dto.BehandlingsAarDto
+import no.nav.k9punsj.journalpost.dto.IdentDto
+import no.nav.k9punsj.journalpost.dto.JournalpostInfoDto
+import no.nav.k9punsj.journalpost.dto.JournalpostMottaksHaandteringDto
+import no.nav.k9punsj.journalpost.dto.KopierJournalpostDto
+import no.nav.k9punsj.journalpost.dto.KopierJournalpostInfo
+import no.nav.k9punsj.journalpost.dto.LukkJournalpostDto
+import no.nav.k9punsj.journalpost.dto.PunsjJournalpost
+import no.nav.k9punsj.journalpost.dto.PunsjJournalpostKildeType
+import no.nav.k9punsj.journalpost.dto.ResultatDto
+import no.nav.k9punsj.journalpost.dto.SettPåVentDto
+import no.nav.k9punsj.journalpost.dto.SkalTilInfotrygdSvar
+import no.nav.k9punsj.journalpost.dto.utledK9sakFagsakYtelseType
+import no.nav.k9punsj.openapi.OasDokumentInfo
+import no.nav.k9punsj.openapi.OasFeil
+import no.nav.k9punsj.openapi.OasJournalpostDto
+import no.nav.k9punsj.openapi.OasJournalpostIder
+import no.nav.k9punsj.tilgangskontroll.AuthenticationHandler
+import no.nav.k9punsj.tilgangskontroll.InnloggetUtils
+import no.nav.k9punsj.tilgangskontroll.abac.IPepClient
+import no.nav.k9punsj.tilgangskontroll.azuregraph.IAzureGraphService
+import no.nav.k9punsj.utils.ServerRequestUtils.hentNorskIdentHeader
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.http.HttpHeaders
+import org.springframework.http.HttpStatus
+import org.springframework.web.reactive.function.BodyExtractors
+import org.springframework.web.reactive.function.server.ServerRequest
+import org.springframework.web.reactive.function.server.ServerResponse
+import org.springframework.web.reactive.function.server.bodyValueAndAwait
+import org.springframework.web.reactive.function.server.buildAndAwait
+import org.springframework.web.reactive.function.server.json
+import java.time.LocalDate
+import java.time.LocalDateTime
+import java.util.*
+import java.util.regex.Pattern
+import kotlin.coroutines.coroutineContext
+
+@Configuration
+internal class JournalpostDriftRoutes(
+    private val authenticationHandler: AuthenticationHandler,
+    private val journalpostService: JournalpostService,
+    private val aksjonspunktService: AksjonspunktService,
+    private val azureGraphService: IAzureGraphService
+) {
+
+    private companion object {
+        private const val JournalpostIdKey = "journalpost_id"
+        private val logger: Logger = LoggerFactory.getLogger(JournalpostDriftRoutes::class.java)
+    }
+
+    internal object Urls {
+        // for drift i prod
+        internal const val ResettInfoOmJournalpost = "/journalpost/resett/{$JournalpostIdKey}"
+        internal const val HentHvaSomHarBlittSendtInn = "/journalpost/hentForDebugg/{$JournalpostIdKey}"
+        internal const val LukkJournalposterDebugg = "/journalpost/lukkDebugg"
+        internal const val LukkJournalpostDebugg = "/journalpost/lukkDebugg/{$JournalpostIdKey}"
+    }
+
+    @Bean
+    fun JournalpostDriftRoutes() = SaksbehandlerRoutes(authenticationHandler) {
+        GET("/api${Urls.ResettInfoOmJournalpost}") { request ->
+            RequestContext(coroutineContext, request) {
+                val journalpostId = request.journalpostId()
+
+                val journalpost = journalpostService.hentHvisJournalpostMedId(journalpostId)
+                    ?: return@RequestContext ServerResponse
+                        .notFound()
+                        .buildAndAwait()
+                val kanSendeInn = journalpostService.kanSendeInn(listOf(journalpostId))
+
+                if (kanSendeInn) {
+                    val nyVerdi = journalpost.copy(skalTilK9 = null)
+                    journalpostService.lagre(nyVerdi)
+
+                    return@RequestContext ServerResponse
+                        .ok()
+                        .json()
+                        .bodyValueAndAwait(ResultatDto("Journalpost med id $journalpostId har blitt resatt!"))
+                }
+                return@RequestContext ServerResponse
+                    .badRequest()
+                    .bodyValueAndAwait("Kan ikke endre på en journalpost som har blitt sendt fra punsj")
+            }
+        }
+
+        GET("/api${Urls.LukkJournalpostDebugg}") { request ->
+            RequestContext(coroutineContext, request) {
+                val journalpostId = request.journalpostId()
+                val enhet = azureGraphService.hentEnhetForInnloggetBruker().trimIndent().take(4)
+
+                journalpostService.hentHvisJournalpostMedId(journalpostId)
+                    ?: return@RequestContext ServerResponse
+                        .notFound()
+                        .buildAndAwait()
+
+                val kanSendeInn = journalpostService.kanSendeInn(listOf(journalpostId))
+                if (kanSendeInn) {
+                    aksjonspunktService.settUtførtPåAltSendLukkOppgaveTilK9Los(journalpostId, false, null)
+                    val (status, body) = journalpostService.settTilFerdig(
+                        journalpostId = journalpostId,
+                        ferdigstillJournalpost = false,
+                        enhet = enhet,
+                        sak = null,
+                        søkerIdentitetsnummer = null
+                    )
+                    if (!status.is2xxSuccessful) {
+                        return@RequestContext ServerResponse.status(status)
+                            .bodyValueAndAwait(body!!)
+                    }
+                    logger.info("Journalpost lukkes", keyValue("journalpost_id", journalpostId))
+
+                    return@RequestContext ServerResponse
+                        .ok()
+                        .json()
+                        .bodyValueAndAwait(ResultatDto("Journalpost med id $journalpostId har blitt lukket i punsj og i los"))
+                } else {
+                    return@RequestContext ServerResponse
+                        .status(HttpStatus.BAD_REQUEST)
+                        .json()
+                        .bodyValueAndAwait(ResultatDto("Journalpost med id $journalpostId har blitt lukket fra før! (Ingen endring)"))
+                }
+            }
+        }
+
+        POST("/api${Urls.LukkJournalposterDebugg}") { request ->
+            RequestContext(coroutineContext, request) {
+                val journalpostIder = request.journalpostIder().journalpostIder.toSet()
+
+                val punsjJper = journalpostService.hentHvisJournalpostMedIder(journalpostIder.toList())
+                if (punsjJper.isEmpty()) {
+                    return@RequestContext ServerResponse
+                        .notFound()
+                        .buildAndAwait()
+                }
+
+                val punsjJperIder = punsjJper.keys.map { it.journalpostId }.toSet()
+                val fantIkkeIPunsjTekst = diffTekst(journalpostIder, punsjJperIder, "Fantes ikke i punsj")
+
+                val uferdigePunsj = punsjJper.filter { !it.value }.map { it.key.journalpostId }.toSet()
+                val alleredeLukketIPunsjTekst = diffTekst(punsjJperIder, uferdigePunsj, "Allerede lukket i punsj")
+                if (uferdigePunsj.isEmpty()) {
+                    return@RequestContext ServerResponse
+                        .status(HttpStatus.BAD_REQUEST)
+                        .json()
+                        .bodyValueAndAwait(
+                            ResultatDto(
+                                "Alle er ferdig behandlet i punsj eller finnes ikke i punsj: " +
+                                    "$alleredeLukketIPunsjTekst $fantIkkeIPunsjTekst"
+                            )
+                        )
+                }
+
+                val medSafStatus =
+                    uferdigePunsj.associateWith { journalpostService.hentSafJournalPost(it)!!.journalstatus }
+                val ferdigStatuser =
+                    arrayOf(SafDtos.Journalstatus.FERDIGSTILT.name, SafDtos.Journalstatus.JOURNALFOERT.name)
+                val ferdigeSaf = medSafStatus.filter { it.value in ferdigStatuser }.keys
+                val ikkeLukketISafTekst =
+                    diffTekst(uferdigePunsj, ferdigeSaf, "Ikke lukket i SAF så disse ble ikke ferdigstilt")
+                if (ferdigeSaf.isEmpty()) {
+                    return@RequestContext ServerResponse
+                        .status(HttpStatus.BAD_REQUEST)
+                        .json()
+                        .bodyValueAndAwait(
+                            ResultatDto(
+                                "Alle er ferdig behandlet i punsj, finnes ikke i punsj eller ikke lukket i SAF. " +
+                                    "$ikkeLukketISafTekst. $fantIkkeIPunsjTekst $alleredeLukketIPunsjTekst"
+                            )
+                        )
+                }
+
+                aksjonspunktService.settUtførtPåAltSendLukkOppgaveTilK9Los(ferdigeSaf, false, null)
+                journalpostService.settAlleTilFerdigBehandlet(ferdigeSaf.toList())
+
+
+                return@RequestContext ServerResponse.status(HttpStatus.OK)
+                    .bodyValueAndAwait(
+                        ResultatDto(
+                            "Lukket journalposter: $ferdigeSaf. $fantIkkeIPunsjTekst $alleredeLukketIPunsjTekst $ikkeLukketISafTekst"
+                        )
+                    )
+            }
+        }
+
+        GET("/api${Urls.HentHvaSomHarBlittSendtInn}") { request ->
+            RequestContext(coroutineContext, request) {
+                val journalpostId = request.journalpostId()
+                val journalpost = journalpostService.hentHvisJournalpostMedId(journalpostId)
+                    ?: return@RequestContext ServerResponse
+                        .notFound()
+                        .buildAndAwait()
+
+                if (journalpost.payload != null) {
+                    try {
+                        journalpostService.hentJournalpostInfo(journalpostId = request.journalpostId())
+                            ?: return@RequestContext ServerResponse
+                                .notFound()
+                                .buildAndAwait()
+
+                        return@RequestContext ServerResponse
+                            .status(HttpStatus.OK)
+                            .json()
+                            .bodyValueAndAwait(journalpost.payload)
+                    } catch (cause: IkkeStøttetJournalpost) {
+                        return@RequestContext ServerResponse
+                            .status(HttpStatus.CONFLICT)
+                            .json()
+                            .bodyValueAndAwait("""{"type":"punsj://ikke-støttet-journalpost"}""")
+                    } catch (case: IkkeTilgang) {
+                        return@RequestContext ServerResponse
+                            .status(HttpStatus.FORBIDDEN)
+                            .buildAndAwait()
+                    }
+                }
+                return@RequestContext ServerResponse
+                    .badRequest()
+                    .buildAndAwait()
+            }
+        }
+    }
+
+    private fun diffTekst(setA: Set<String>, setB: Set<String>, prefix: String): String {
+        val diff = setA.minus(setB)
+        return if (diff.isNotEmpty()) "$prefix: $diff" else ""
+    }
+
+    private fun ServerRequest.journalpostId(): String = pathVariable(JournalpostIdKey)
+
+    internal data class JournalpostIderRequest(val journalpostIder: List<String>)
+
+    private suspend fun ServerRequest.journalpostIder(): JournalpostIderRequest =
+        body(BodyExtractors.toMono(JournalpostIderRequest::class.java)).awaitFirst()
+
+}
