+++ src/main/kotlin/no/nav/k9punsj/journalpost/JournalpostService.kt
@@ -51,13 +51,6 @@ class JournalpostService(
         safGateway.hentJournalpostInfo(journalpostId)
 
     @Deprecated("Periode skal hentes fra søknaden i soknadservice. FIXME")
-    internal suspend fun hentBehandlingsAar(journalpostId: String): Int {
-        val behandlingsAar = journalpostRepository.hent(journalpostId).behandlingsAar
-        logger.info("Hentet behandlingsår ($behandlingsAar) for journalpost: $journalpostId")
-        return behandlingsAar ?: LocalDate.now().year
-    }
-
-    @Deprecated("Periode skal hentes fra søknaden i soknadservice. FIXME")
     internal suspend fun lagreBehandlingsAar(journalpostId: String, behandlingsAar: Int) {
         val journalpost = journalpostRepository.hentHvis(journalpostId)
         if (journalpost != null) {
