+++ src/main/kotlin/no/nav/k9punsj/omsorgspengerutbetaling/OmsorgspengerutbetalingService.kt
@@ -64,7 +64,7 @@ internal class OmsorgspengerutbetalingService(
     }
 
     internal suspend fun henteSøknad(søknadId: String): ServerResponse {
-        val søknad = mappeService.hentSøknad(søknadId)
+        val søknad = soknadService.hentSøknad(søknadId)
 
         return if (søknad == null) {
             ServerResponse
@@ -112,7 +112,7 @@ internal class OmsorgspengerutbetalingService(
     }
 
     internal suspend fun sendEksisterendeSøknad(sendSøknad: SendSøknad): ServerResponse {
-        val søknadEntitet = mappeService.hentSøknad(sendSøknad.soeknadId)
+        val søknadEntitet = soknadService.hentSøknad(sendSøknad.soeknadId)
             ?: return ServerResponse.badRequest().buildAndAwait()
 
         try {
@@ -204,7 +204,7 @@ internal class OmsorgspengerutbetalingService(
     }
 
     internal suspend fun validerSøknad(soknadTilValidering: OmsorgspengerutbetalingSøknadDto): ServerResponse {
-        val søknadEntitet = mappeService.hentSøknad(soknadTilValidering.soeknadId)
+        val søknadEntitet = soknadService.hentSøknad(soknadTilValidering.soeknadId)
             ?: return ServerResponse
                 .badRequest()
                 .buildAndAwait()
