+++ src/main/kotlin/no/nav/k9punsj/pleiepengersyktbarn/PleiepengerSyktBarnService.kt
@@ -65,7 +65,7 @@ internal class PleiepengerSyktBarnService(
     }
 
     internal suspend fun henteSøknad(søknadId: String): ServerResponse {
-        val søknad = mappeService.hentSøknad(søknadId)
+        val søknad = soknadService.hentSøknad(søknadId)
             ?: return ServerResponse
                 .notFound()
                 .buildAndAwait()
@@ -101,7 +101,7 @@ internal class PleiepengerSyktBarnService(
     }
 
     internal suspend fun sendEksisterendeSøknad(sendSøknad: SendSøknad): ServerResponse {
-        val søknadEntitet = mappeService.hentSøknad(sendSøknad.soeknadId)
+        val søknadEntitet = soknadService.hentSøknad(sendSøknad.soeknadId)
             ?: return ServerResponse.badRequest().buildAndAwait()
 
         try {
@@ -198,7 +198,7 @@ internal class PleiepengerSyktBarnService(
     }
 
     internal suspend fun validerSøknad(soknadTilValidering: PleiepengerSyktBarnSøknadDto): ServerResponse {
-        val søknadEntitet = mappeService.hentSøknad(soknadTilValidering.soeknadId)
+        val søknadEntitet = soknadService.hentSøknad(soknadTilValidering.soeknadId)
             ?: return ServerResponse
                 .badRequest()
                 .buildAndAwait()
